Project completetion notes
An image of the finished dish
Under the description, add an “Ingredients” heading followed by an unordered list
Under the image, it should have an appropriately sized “Description” heading followed by a paragraph or two describing the recipe
Under the description, add an “Ingredients” heading followed by an unordered list of the ingredients needed for the recipe
Finally, under the ingredients list, add a “Steps” heading followed by an ordered list of the steps needed for making the dish
Add two more recipes with identical page structures to the recipe page you’ve already created
Don’t forget to link to the new recipes on the index page. Also, consider putting all the links in an unordered list so they aren’t all on one line.